other:重构设置页面,关系组页面移动到设置主页, delete:删除多余页面
