+++ lib/pages/setting/attribute_info/state.dart
@@ -3,7 +3,7 @@
 import 'package:get/get.dart';
 import 'package:orginone/dart/base/schema.dart';
 import 'package:orginone/dart/core/getx/base_get_state.dart';
-import 'package:orginone/pages/setting/home/setting/state.dart';
+import 'package:orginone/pages/setting/home/state.dart';
 
 class AttributeInfoState extends BaseGetState{
