+++ lib/pages/setting/dict_info/state.dart
@@ -4,7 +4,7 @@
 import 'package:get/get.dart';
 import 'package:orginone/dart/base/schema.dart';
 import 'package:orginone/dart/core/getx/base_get_state.dart';
-import 'package:orginone/pages/setting/home/setting/state.dart';
+import 'package:orginone/pages/setting/home/state.dart';
 
 class DictInfoState extends BaseGetState{
   late SettingNavModel data;
