--- lib/pages/setting/home/setting/binding.dart
@@ -1,12 +1,0 @@
-import 'package:get/get.dart';
-import 'package:orginone/dart/core/getx/base_bindings.dart';
-
-import 'logic.dart';
-
-class SettingFunctionBinding extends BaseBindings<SettingFunctionController> {
-  @override
-  SettingFunctionController getController() {
-   return SettingFunctionController();
-  }
-
-}
