--- lib/pages/setting/relationship_group/binding.dart
@@ -1,14 +1,0 @@
-import 'package:get/get.dart';
-import 'package:orginone/dart/core/getx/base_bindings.dart';
-
-import 'logic.dart';
-
-
-
-class RelationGroupBinding extends BaseBindings<RelationGroupController> {
-  @override
-  RelationGroupController getController() {
-   return RelationGroupController();
-  }
-
-}
