+++ lib/routers.dart
@@ -29,7 +29,6 @@ import 'package:orginone/pages/work/work_start/view.dart';
 import 'package:orginone/pages/setting/contact_page.dart';
 import 'package:orginone/pages/setting/dict_info/view.dart';
 import 'package:orginone/pages/setting/home/binding.dart';
-import 'package:orginone/pages/setting/home/setting/binding.dart';
 import 'package:orginone/pages/setting/home/view.dart';
 import 'package:orginone/pages/setting/mine_unit_page.dart';
 import 'package:orginone/pages/setting/person/cardbag/bindings.dart';
@@ -90,13 +89,10 @@ import 'pages/setting/company_info/view.dart';
 import 'pages/setting/department_info/binding.dart';
 import 'pages/setting/department_info/view.dart';
 import 'pages/setting/dict_info/binding.dart';
-import 'pages/setting/home/setting/view.dart';
 import 'pages/setting/out_agency_info/binding.dart';
 import 'pages/setting/out_agency_info/view.dart';
 import 'pages/setting/permission_info/binding.dart';
 import 'pages/setting/permission_info/view.dart';
-import 'pages/setting/relationship_group/binding.dart';
-import 'pages/setting/relationship_group/view.dart';
 import 'pages/setting/role_settings/binding.dart';
 import 'pages/setting/role_settings/view.dart';
 import 'pages/setting/station_info/binding.dart';
@@ -144,7 +140,6 @@ class Routers {
 
   // 设置
   static const String settingCenter = "/settingCenter";
-  static const String settingFunction = "/settingFunction";
   static const String mineUnit = "/mineUnit";
   static const String contact = "/contact";
   static const String cohorts = "/cohorts";
@@ -242,9 +237,6 @@ class Routers {
 
   //文件夹
   static const String file = '/file';
-
-  //关系列表
-  static const String relationGroup = '/relationGroup';
 
   //内部机构详情
   static const String departmentInfo = '/departmentInfo';
@@ -476,11 +468,6 @@ class Routers {
         binding: SettingCenterBinding(),
       ),
       GetPage(
-        name: Routers.settingFunction,
-        page: () => SettingFunctionPage(),
-        binding: SettingFunctionBinding(),
-      ),
-      GetPage(
         name: Routers.companyInfo,
         page: () => CompanyInfoPage(),
         binding: CompanyInfoBinding(),
@@ -489,11 +476,6 @@ class Routers {
         name: Routers.version,
         page: () => const VersionPage(),
         binding: VersionBinding(),
-      ),
-      GetPage(
-        name: Routers.relationGroup,
-        page: () => RelationGroupPage(),
-        binding: RelationGroupBinding(),
       ),
       GetPage(
         name: Routers.departmentInfo,
